Crate idmage

This is the entry point, where an HTML request is parsed and routed.

Modules

agent
Experts that operate the canvas

builder
Creates the SVG from a canvas

canvas
Holds all the current information about the artwork and current status.

controller
The controller allows agents to either update their current command, or execute their precious.

graphic
A single or grouped element that caa be placed on a canvas.

profiles
Different visual identities that can be interpretated.

svgpower
Some common SVG elements and attributes

Statics

static_rocket_route_info_for_canvas
Rocket code generated static route information structure.

static_rocket_route_info_for_create
Rocket code generated static route information structure.

static_rocket_route_info_for_generate
Rocket code generated static route information structure.

static_rocket_route_info_for_index
Rocket code generated static route information structure.

Functions

canvas
Individual profiles, for setting the style

create

generate

index

main

rocket_route_fn_canvas

rocket_route_fn_create

rocket_route_fn_generate

rocket_route_fn_index